Shim gfxdraw when possible
